refactor : port 위치 port 디렉토리로 변경
