+++ src/main/java/com/example/bak/community/application/command/CommunityCommandService.java
@@ -1,5 +1,7 @@
 package com.example.bak.community.application.command;
 
+import com.example.bak.community.application.command.port.CommunityCommandPort;
+import com.example.bak.community.application.command.port.CompanyValidationPort;
 import com.example.bak.community.application.dto.CommunityResult;
 import com.example.bak.community.domain.Community;
 import com.example.bak.global.exception.BusinessException;
