+++ src/main/java/com/example/bak/community/application/command/port/CompanyValidationPort.java
@@ -1,4 +1,4 @@
-package com.example.bak.community.application.command;
+package com.example.bak.community.application.command.port;
 
 public interface CompanyValidationPort {
